refactor: separated render and list on StockListCard
