+++ src/main/java/edu/ntnu/idatt2003/millions/view/exchange/stocks/StocksView.java
@@ -1,10 +1,10 @@
 package edu.ntnu.idatt2003.millions.view.exchange.stocks;
 
 import edu.ntnu.idatt2003.millions.controller.PortfolioController;
-import edu.ntnu.idatt2003.millions.observer.GameObserver;
 import edu.ntnu.idatt2003.millions.service.GameService;
 import edu.ntnu.idatt2003.millions.util.LanguageManager;
 import edu.ntnu.idatt2003.millions.view.component.AvailableFundsCard;
+import edu.ntnu.idatt2003.millions.view.component.Pagination;
 import edu.ntnu.idatt2003.millions.view.component.PortfolioValueCard;
 import edu.ntnu.idatt2003.millions.view.component.SearchBar;
 import edu.ntnu.idatt2003.millions.view.component.StyledText;
@@ -29,31 +29,24 @@ import java.text.MessageFormat;
  *
  * <p>Search is triggered explicitly by pressing Enter or clicking the search button,
  * delegated entirely to {@link SearchBar}. Sort state is managed by
- * {@link StocksSort} inside the table. The clear sort button is shown only when a
- * sort is active and is placed right-aligned on the same row as the search bar via
- * {@link StocksListCard#setOnSortChanged(Runnable)}.
+ * {@link StocksSort} inside the table.
  */
-public class StocksView extends VBox implements GameObserver {
+public class StocksView extends VBox {
 
     private final StocksListCard stocksListCard;
-    private final SearchBar searchBar;
-    private final GameService gameService;
     private final StyledText statusLabel = StyledText.widgetLabel();
+    private final Pagination pagination;
 
     /**
      * Constructs a new StocksView.
-     *
-     * <p>Wires the search bar to {@link StocksListCard#filter(String)} and registers
-     * a sort-change callback so the clear sort button visibility and the status
-     * label are kept in sync with the table state.
      *
      * @param gameService the game service containing player and exchange state
      * @param controller  the controller used to open buy/sell dialogs
      */
     public StocksView(GameService gameService, PortfolioController controller) {
-        this.gameService = gameService;
         this.stocksListCard = new StocksListCard(gameService, controller);
-        this.searchBar = new SearchBar(
+        this.pagination = new Pagination(StocksListCard.PAGE_SIZE, stocksListCard::setPage);
+        SearchBar searchBar = new SearchBar(
                 "exchange.stocks.search.placeholder",
                 "search.button",
                 term -> {
@@ -62,12 +55,13 @@ public class StocksView extends VBox implements GameObserver {
                 });
 
         Button clearSortButton = new Button(LanguageManager.get("exchange.stocks.sort.clear"));
-        clearSortButton.getStyleClass().add("search-clear-button");
+        clearSortButton.getStyleClass().add("clear-sort-button");
         clearSortButton.setOpacity(0);
         clearSortButton.setOnAction(e -> stocksListCard.clearSort());
 
-        stocksListCard.setOnSortChanged(() -> {
+        stocksListCard.setOnRefreshed(() -> {
             clearSortButton.setOpacity(stocksListCard.isSortActive() ? 1 : 0);
+            pagination.update(stocksListCard.getCurrentPage(), stocksListCard.getFilteredCount());
             updateStatus();
         });
 
@@ -91,9 +85,8 @@ public class StocksView extends VBox implements GameObserver {
                         "exchange.stocks.unrealized.sub"))
         );
 
-        gameService.addObserver(this);
-
-        getChildren().addAll(cards, searchRow, statusLabel, stocksListCard);
+        getChildren().addAll(cards, searchRow, statusLabel, pagination, stocksListCard);
+        updateStatus();
     }
 
     /**
@@ -105,15 +98,6 @@ public class StocksView extends VBox implements GameObserver {
                 LanguageManager.get("exchange.stocks.status"),
                 stocksListCard.getFilteredCount(),
                 stocksListCard.getTotalCount()));
-    }
-
-    /**
-     * Called when the game state changes. Refreshes the status label to reflect
-     * any change in the total or filtered stock count.
-     */
-    @Override
-    public void onGameUpdated() {
-        updateStatus();
     }
 
     /**
