+++ src/main/resources/css/buttons.css
@@ -1,4 +1,4 @@
-.clear-button {
+.clear-sort-button {
     -fx-min-height: 35;
     -fx-pref-height: 35;
     -fx-max-height: 35;
@@ -8,10 +8,14 @@
     -fx-font-weight: 500;
     -fx-padding: 0 22 0 22;
     -fx-cursor: hand;
-    -fx-border-color: #EDEDED;
     -fx-border-radius: 10;
     -fx-background-color: white;
+    -fx-border-color: #EDEDED;
     -fx-text-fill: black;
+}
+.clear-sort-button:hover {
+    -fx-background-color: -primary-dark;
+    -fx-text-fill: white;
 }
 
  .pagination-button {
