+++ src/main/resources/i18n/messages_en.properties
@@ -102,6 +102,7 @@ exchange.stocks.no.data=—
 # Pagination
 pagination.prev= < Previous
 pagination.next=Next >
+pagination.pages={0} / {1} pages
 # Exit dialog
 exit.confirmHeader=Are you sure you want to exit?
 exit.confirmContent=You can save the game before exiting.
