+++ src/main/resources/i18n/messages_no.properties
@@ -17,24 +17,24 @@ start.loadButton=Last inn spill
 start.new.currencyLabel=Velg valuta for filen
 # Navbar
 nav.mySides=Mine sider
-nav.exchange=Børs
+nav.exchange=Bï¿½rs
 nav.leaderboard=Ledertavle
 nav.help=Hjelp
 nav.saveGame=Lagre spill
 nav.exitGame=Avslutt spill
 # Searchbar
-search.button=Søk
+search.button=Sï¿½k
 # Dashboard
 dashboard.title=Mine sider
-dashboard.tab.portfolio=Portefølje
+dashboard.tab.portfolio=Portefï¿½lje
 dashboard.tab.transactions=Transaksjoner
-dashboard.tab.watchlist=Overvøkningsliste
-dashboard.tab.loans=Løn
+dashboard.tab.watchlist=Overvï¿½kningsliste
+dashboard.tab.loans=Lï¿½n
 # Dashboard - Portfolio tab
 dashboard.netWorth=Nettoverdi
 dashboard.weeklyChange=Utvikling denne uken
 dashboard.availableFunds=Tilgjengelig for handel
-dashboard.portfolioValue=Porteføljeverdi
+dashboard.portfolioValue=Portefï¿½ljeverdi
 dashboard.status=Status
 dashboard.portfolio.title=Beholdning
 dashboard.portfolio.company=Selskap
@@ -47,7 +47,7 @@ dashboard.portfolio.valueNok=Verdi NOK
 dashboard.portfolio.returnPct=Avkast. %
 dashboard.portfolio.returnNok=Avkast. NOK
 dashboard.portfolio.total=Totalt
-dashboard.portfolio.buy=Kjøp
+dashboard.portfolio.buy=Kjï¿½p
 dashboard.portfolio.sell=Selg
 dashboard.portfolio.sellAll=Selg alt
 dashboard.exploreStocks=Utforsk andre aksjer
@@ -58,11 +58,11 @@ dashboard.realized.net=Netto realisert
 dashboard.realized.tax=Total skatt betalt
 dashboard.realized.commission=Total kurtasje betalt
 dashboard.realized.count=Antall salg
-dashboard.realized.empty=Selg aksjer for å se realisert avkastning her.
+dashboard.realized.empty=Selg aksjer for ï¿½ se realisert avkastning her.
 dashboard.sinceStart=siden start
-dashboard.portfolio.empty=Du eier ingen andeler ennø
+dashboard.portfolio.empty=Du eier ingen andeler ennï¿½
 # Exchange
-exchange.title=Børs
+exchange.title=Bï¿½rs
 exchange.tab.overview=Oversikt
 exchange.tab.stocks=Aksjer
 exchange.tab.analysis=Analyser
@@ -76,15 +76,15 @@ exchange.overview.fellThisWeek=Sank denne uken (-)
 exchange.overview.weeklyWinners=Ukens vinnere
 exchange.overview.weeklyLosers=Ukens tapere
 # Exchange - Stocks tab
-exchange.stocks.portfolio=Din portefølje
+exchange.stocks.portfolio=Din portefï¿½lje
 exchange.stocks.portfolio.sub=Startkapital
 exchange.stocks.available=Tilgjengelig
-exchange.stocks.available.sub=Etter kjøp
+exchange.stocks.available.sub=Etter kjï¿½p
 exchange.stocks.invested=Investert
 exchange.stocks.invested.positions={0} posisjoner
 exchange.stocks.unrealized=Urealisert gevinst
-exchange.stocks.unrealized.sub=Siden kjøp
-exchange.stocks.search.placeholder=Søk på ticker eller selskapet...
+exchange.stocks.unrealized.sub=Siden kjï¿½p
+exchange.stocks.search.placeholder=Sï¿½k pï¿½ ticker eller selskapet...
 exchange.stocks.sort.clear=Nullstill
 exchange.stocks.status=Viser {0} av {1} aksjer. Kursdata oppdateres ved ukeslutt
 exchange.stocks.col.ticker=Ticker
@@ -96,24 +96,25 @@ exchange.stocks.col.changePct=+/- %
 exchange.stocks.col.trend=Trend
 exchange.stocks.col.trade=Handel
 exchange.stocks.badge.owner=Eier
-exchange.stocks.buy=Kjøp
+exchange.stocks.buy=Kjï¿½p
 exchange.stocks.sell=Selg
 exchange.stocks.no.data=0
 # Pagination
 pagination.prev=< Forrige
 pagination.next=Neste >
+pagination.pages={0} / {1} sider
 # Exit dialog
-exit.confirmHeader=Er du sikker på at du vil avslutte?
-exit.confirmContent=Du kan lagre spillet før du avslutter.
+exit.confirmHeader=Er du sikker pï¿½ at du vil avslutte?
+exit.confirmContent=Du kan lagre spillet fï¿½r du avslutter.
 exit.saveAndExit=Lagre og avslutt
-exit.exitWithoutSaving=Avslutt uten å lagre
+exit.exitWithoutSaving=Avslutt uten ï¿½ lagre
 exit.cancel=Avbryt
 # Player status
 status.novice=Nybegynner
 status.investor=Investor
 status.speculator=Spekulant
 # Dashboard - Portfolio dialogs
-dialog.buy.title=Kjøp aksjer
+dialog.buy.title=Kjï¿½p aksjer
 dialog.sell.title=Selg aksjer
 dialog.sellAll.title=Selg alt
 dialog.stock.label=Aksje
@@ -130,19 +131,19 @@ dialog.summary.commissionSell=Kurtasje (1%)
 dialog.summary.tax=Skatt (30% av gevinst)
 dialog.summary.totalCost=Totalkostnad
 dialog.summary.totalReceived=Du mottar
-dialog.balance.available=Tilgjengelig nø
-dialog.balance.afterBuy=Etter kjøp
+dialog.balance.available=Tilgjengelig nï¿½
+dialog.balance.afterBuy=Etter kjï¿½p
 dialog.balance.afterSell=Etter salg
-dialog.error.insufficientFunds=Du har ikke nok penger til dette kjøpet.
-dialog.profit.gain=Gevinst på dette salget
-dialog.profit.loss=Tap på dette salget
+dialog.error.insufficientFunds=Du har ikke nok penger til dette kjï¿½pet.
+dialog.profit.gain=Gevinst pï¿½ dette salget
+dialog.profit.loss=Tap pï¿½ dette salget
 dialog.button.cancel=Avbryt
-dialog.button.confirmBuy=Bekreft kjøp
+dialog.button.confirmBuy=Bekreft kjï¿½p
 dialog.button.confirmSell=Bekreft salg
 dialog.button.sellAll=Selg alt
 # Dashboard - Portfolio receipts
-receipt.buy.title=Kjøp gjennomført
-receipt.sell.title=Salg gjennomført
+receipt.buy.title=Kjï¿½p gjennomfï¿½rt
+receipt.sell.title=Salg gjennomfï¿½rt
 receipt.stock.label=Aksje
 receipt.meta.quantity=Antall
 receipt.meta.price=Pris per andel
@@ -154,8 +155,8 @@ receipt.summary.commissionSell=Kurtasje (1%)
 receipt.summary.tax=Skatt (30%)
 receipt.summary.totalCost=Totalkostnad
 receipt.summary.totalReceived=Du mottok
-receipt.profit.gain=Gevinst på salget
-receipt.profit.loss=Tap på salget
-receipt.balance.before=Saldo før
+receipt.profit.gain=Gevinst pï¿½ salget
+receipt.profit.loss=Tap pï¿½ salget
+receipt.balance.before=Saldo fï¿½r
 receipt.balance.after=Saldo etter
 receipt.button.close=Lukk
